Move messages to separate directory
