+++ arangod/Pregel/Conductor/States/ComputingState.cpp
@@ -5,7 +5,6 @@
 #include "Pregel/Conductor/States/State.h"
 #include "Pregel/MasterContext.h"
 #include "Pregel/PregelFeature.h"
-#include "Pregel/WorkerConductorMessages.h"
 #include "velocypack/Builder.h"
 #include "velocypack/Iterator.h"
