+++ arangod/Pregel/Conductor/WorkerApi.h
@@ -25,7 +25,8 @@
 #include "Cluster/ClusterTypes.h"
 #include "Futures/Future.h"
 #include "Pregel/ExecutionNumber.h"
-#include "Pregel/WorkerConductorMessages.h"
+#include "Pregel/Messaging/WorkerMessages.h"
+#include "Pregel/Messaging/ConductorMessages.h"
 #include "Pregel/WorkerInterface.h"
 #include "Utils/DatabaseGuard.h"
 #include "Pregel/Connection/Connection.h"
